Move temperature profiles to its own module
