+++ src/ClimateMachine.jl
@@ -11,6 +11,7 @@ include(joinpath("InputOutput", "Writers", "Writers.jl"))
 include(joinpath("Common", "ConfigTypes", "ConfigTypes.jl"))
 include(joinpath("Utilities", "VariableTemplates", "VariableTemplates.jl"))
 include(joinpath("Common", "MoistThermodynamics", "MoistThermodynamics.jl"))
+include(joinpath("Atmos", "TemperatureProfiles", "TemperatureProfiles.jl"))
 include(joinpath(
     "Atmos",
     "Parameterizations",
