+++ test/Numerics/DGmethods/compressible_Navier_Stokes/ref_state.jl
@@ -10,6 +10,7 @@ using ClimateMachine.ODESolvers
 using ClimateMachine.GenericCallbacks
 using ClimateMachine.Atmos
 using ClimateMachine.VariableTemplates
+using ClimateMachine.TemperatureProfiles
 using ClimateMachine.MoistThermodynamics
 using LinearAlgebra
 using StaticArrays
